requirements-app.txt
aiohttp==3.6.2
fastapi[all]==0.45.0
pyppeteer==0.0.25
Pillow==6.2.1
google-cloud-storage==1.23.0
google-cloud-pubsub==1.1.0
tzlocal==2.0.0